create game_state, player, life, menus and archtecture of folders
