+++ __main__.py
@@ -1,4 +1,4 @@
-from entities.game_state import Crew
+from controller.game_state import Crew
 
 if __name__ == '__main__':
     game = Crew()
