+++ entities/player_entity.py
@@ -1,6 +1,16 @@
-import pygame
+from pygame import *
 
 
-class Player(pygame.sprite.Sprite):
-    def __init__():
-        pass
+class Player(sprite.Sprite):
+    def __init__(self):
+        sprite.Sprite.__init__(self)
+        self.image = IMAGES['player']
+        self.rect = self.image.get_rect(topleft=(375, 540))
+        self.speed = 5
+
+    def update(self, keys, *args):
+        if keys[K_LEFT] and self.rect.x > 10:
+            self.rect.x -= self.speed
+        if keys[K_RIGHT] and self.rect.x < 740:
+            self.rect.x += self.speed
+        game.screen.blit(self.image, self.rect)
