+++ entities/text_entity.py
@@ -1,3 +1,6 @@
+from pygame import *
+
+
 class Text(object):
     def __init__(self, textFont, size, message, color, xpos, ypos):
         self.font = font.Font(textFont, size)
