+++ paths.py
@@ -1,0 +1,7 @@
+
+from os.path import abspath, dirname
+
+
+BASE_PATH = abspath(dirname(__file__))
+FONT_PATH = BASE_PATH + '/fonts/'
+IMAGE_PATH = BASE_PATH + '/assets/'
